pyproject.toml
[project]
name = "object-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "pandas>=2.3.0",
    "pillow>=11.2.1",
    "tensorboard>=2.19.0",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
]
